+++ src/FontAwesome5.Net/Converters/ImageSourceSvgConverter.cs
@@ -1,0 +1,44 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Globalization;
+using System.Linq;
+using System.Text;
+using System.Windows.Data;
+using System.Windows.Markup;
+using System.Windows.Media;
+
+using FontAwesome5.Extensions;
+
+namespace FontAwesome5.Converters
+{
+    /// <summary>
+    /// Converts a FontAwesomIcon to an ImageSource. Use the ConverterParameter to pass the Brush.
+    /// </summary>
+    public class ImageSourceSvgConverter : MarkupExtension, IValueConverter
+    {
+        public override object ProvideValue(IServiceProvider serviceProvider)
+        {
+            return this;
+        }
+
+        public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
+        {
+            if (value is not EFontAwesomeIcon)
+            {
+                return null;
+            }
+
+            if (parameter is not Brush brush)
+            {
+                brush = Brushes.Black;
+            }
+
+            return ((EFontAwesomeIcon)value).CreateImageSource(brush, new Pen());
+        }
+
+        public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture)
+        {
+            throw new NotImplementedException();
+        }
+    }
+}