+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 2.1.3
+- Added ImageSourceSvgConverter
+- Made the Font Families adjustable on the static Fonts class
+- Introduced a new ReosurceDictionary (FontAwesome5Dictionary.xaml) to prevent memory leaks
 # 2.1.2
 - Fixed issue #31
 # 2.1.1